requirements.txt
PySide6>=6.4.0
cryptography>=40.0.0
argon2-cffi>=21.0.0
pyotp>=2.8.0
qrcode>=7.0.0
reportlab>=4.0.0
opencv-python>=4.7.0
pillow>=10.0.0
requests>=2.31.0
